requirements.txt
celery==4.0.0
Flask==0.11.1
flower==0.9.1
honcho==0.7.1
suds-jurko==0.6
Flask-JWT==0.3.2
suds-jurko==0.6
Flask-JWT-Extended==1.1.0
flask-marshmallow==0.7.0
SQLAlchemy==1.1.4
Flask-SQLAlchemy==2.1
bagit==1.5.4